telegram-bot/requirements.txt
python-telegram-bot==21.3
flake8==7.1.0
